管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -42,28 +42,32 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
-        //指定されたページ数の一覧画面に表示するデータを取得
-        int page = getPage();
-        List<EmployeeView> employees = service.getPerPage(page);
+        //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-        //全ての従業員データの件数を取得
-        long employeeCount = service.countAll();
+            //指定されたページ数の一覧画面に表示するデータを取得
+            int page = getPage();
+            List<EmployeeView> employees = service.getPerPage(page);
 
-        putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
-        putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
-        putRequestScope(AttributeConst.PAGE, page); //ページ数
-        putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
+            //全ての従業員データの件数を取得
+            long employeeCount = service.countAll();
 
-        //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
-        String flush = getSessionScope(AttributeConst.FLUSH);
-        if (flush != null) {
-            putRequestScope(AttributeConst.FLUSH, flush);
-            removeSessionScope(AttributeConst.FLUSH);
+            putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
+            putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
+            putRequestScope(AttributeConst.PAGE, page); //ページ数
+            putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
+
+            //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
+            String flush = getSessionScope(AttributeConst.FLUSH);
+            if (flush != null) {
+                putRequestScope(AttributeConst.FLUSH, flush);
+                removeSessionScope(AttributeConst.FLUSH);
+            }
+
+            //一覧画面を表示
+            forward(ForwardConst.FW_EMP_INDEX);
+
         }
-
-        //一覧画面を表示
-        forward(ForwardConst.FW_EMP_INDEX);
-
     }
 
     /**
@@ -73,11 +77,15 @@ public class EmployeeAction extends ActionBase {
      */
     public void entryNew() throws ServletException, IOException {
 
-        putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
+      //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-        //新規登録画面を表示
-        forward(ForwardConst.FW_EMP_NEW);
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
+
+            //新規登録画面を表示
+            forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /**
@@ -87,46 +95,50 @@ public class EmployeeAction extends ActionBase {
      */
     public void create() throws ServletException, IOException {
 
-        //CSRF対策 tokenのチェック
-        if (checkToken()) {
+      //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-            //パラメータの値を元に従業員情報のインスタンスを作成する
-            EmployeeView ev = new EmployeeView(
-                    null,
-                    getRequestParam(AttributeConst.EMP_CODE),
-                    getRequestParam(AttributeConst.EMP_NAME),
-                    getRequestParam(AttributeConst.EMP_PASS),
-                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-                    null,
-                    null,
-                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
+            //CSRF対策 tokenのチェック
+            if (checkToken()) {
 
-            //アプリケーションスコープからpepper文字列を取得
-            String pepper = getContextScope(PropertyConst.PEPPER);
+                //パラメータの値を元に従業員情報のインスタンスを作成する
+                EmployeeView ev = new EmployeeView(
+                        null,
+                        getRequestParam(AttributeConst.EMP_CODE),
+                        getRequestParam(AttributeConst.EMP_NAME),
+                        getRequestParam(AttributeConst.EMP_PASS),
+                        toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                        null,
+                        null,
+                        AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-            //従業員情報登録
-            List<String> errors = service.create(ev, pepper);
+                //アプリケーションスコープからpepper文字列を取得
+                String pepper = getContextScope(PropertyConst.PEPPER);
 
-            if (errors.size() > 0) {
-                //登録中にエラーがあった場合
+                //従業員情報登録
+                List<String> errors = service.create(ev, pepper);
 
-                putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-                putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
-                putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
+                if (errors.size() > 0) {
+                    //登録中にエラーがあった場合
 
-                //新規登録画面を再表示
-                forward(ForwardConst.FW_EMP_NEW);
+                    putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+                    putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
+                    putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
 
-            } else {
-                //登録中にエラーがなかった場合
+                    //新規登録画面を再表示
+                    forward(ForwardConst.FW_EMP_NEW);
 
-                //セッションに登録完了のフラッシュメッセージを設定
-                putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
+                } else {
+                    //登録中にエラーがなかった場合
 
-                //一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                    //セッションに登録完了のフラッシュメッセージを設定
+                    putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
+
+                    //一覧画面にリダイレクト
+                    redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                }
+
             }
-
         }
     }
 
@@ -137,20 +149,24 @@ public class EmployeeAction extends ActionBase {
      */
     public void show() throws ServletException, IOException {
 
-        //idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+      //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-        if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            //データが取得できなかった、または論理削除されている場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+
+            //詳細画面を表示
+            forward(ForwardConst.FW_EMP_SHOW);
         }
-
-        putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
-
-        //詳細画面を表示
-        forward(ForwardConst.FW_EMP_SHOW);
     }
 
     /**
@@ -160,22 +176,26 @@ public class EmployeeAction extends ActionBase {
      */
     public void edit() throws ServletException, IOException {
 
-        //idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+      //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-        if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            //データが取得できなかった、または論理削除されている場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+
+            //編集画面を表示する
+            forward(ForwardConst.FW_EMP_EDIT);
+
         }
-
-        putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-        putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
-
-        //編集画面を表示する
-        forward(ForwardConst.FW_EMP_EDIT);
-
     }
 
     /**
@@ -185,42 +205,46 @@ public class EmployeeAction extends ActionBase {
      */
     public void update() throws ServletException, IOException {
 
-        //CSRF対策 tokenのチェック
-        if (checkToken()) {
-            //パラメータの値を元に従業員情報のインスタンスを作成する
-            EmployeeView ev = new EmployeeView(
-                    toNumber(getRequestParam(AttributeConst.EMP_ID)),
-                    getRequestParam(AttributeConst.EMP_CODE),
-                    getRequestParam(AttributeConst.EMP_NAME),
-                    getRequestParam(AttributeConst.EMP_PASS),
-                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-                    null,
-                    null,
-                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
+      //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-            //アプリケーションスコープからpepper文字列を取得
-            String pepper = getContextScope(PropertyConst.PEPPER);
+            //CSRF対策 tokenのチェック
+            if (checkToken()) {
+                //パラメータの値を元に従業員情報のインスタンスを作成する
+                EmployeeView ev = new EmployeeView(
+                        toNumber(getRequestParam(AttributeConst.EMP_ID)),
+                        getRequestParam(AttributeConst.EMP_CODE),
+                        getRequestParam(AttributeConst.EMP_NAME),
+                        getRequestParam(AttributeConst.EMP_PASS),
+                        toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                        null,
+                        null,
+                        AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-            //従業員情報更新
-            List<String> errors = service.update(ev, pepper);
+                //アプリケーションスコープからpepper文字列を取得
+                String pepper = getContextScope(PropertyConst.PEPPER);
 
-            if (errors.size() > 0) {
-                //更新中にエラーが発生した場合
+                //従業員情報更新
+                List<String> errors = service.update(ev, pepper);
 
-                putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-                putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
-                putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
+                if (errors.size() > 0) {
+                    //更新中にエラーが発生した場合
 
-                //編集画面を再表示
-                forward(ForwardConst.FW_EMP_EDIT);
-            } else {
-                //更新中にエラーがなかった場合
+                    putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+                    putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
+                    putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
 
-                //セッションに更新完了のフラッシュメッセージを設定
-                putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
+                    //編集画面を再表示
+                    forward(ForwardConst.FW_EMP_EDIT);
+                } else {
+                    //更新中にエラーがなかった場合
 
-                //一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                    //セッションに更新完了のフラッシュメッセージを設定
+                    putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
+
+                    //一覧画面にリダイレクト
+                    redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                }
             }
         }
     }
@@ -232,18 +256,48 @@ public class EmployeeAction extends ActionBase {
      */
     public void destroy() throws ServletException, IOException {
 
-        //CSRF対策 tokenのチェック
-        if (checkToken()) {
+      //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-            //idを条件に従業員データを論理削除する
-            service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+            //CSRF対策 tokenのチェック
+            if (checkToken()) {
 
-            //セッションに削除完了のフラッシュメッセージを設定
-            putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+                //idを条件に従業員データを論理削除する
+                service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            //一覧画面にリダイレクト
-            redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                //セッションに削除完了のフラッシュメッセージを設定
+                putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+
+                //一覧画面にリダイレクト
+                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+            }
         }
     }
+
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+
+            return true;
+        }
+
+    }
+
 
 }
